Add styles and error checking for project design
Styles were added to match closer to the active
and default state of the desktop version of the
challenge project. In the JS files event listeners
and checks were added to add invalid classes
to display and adjust the necessary styles, icons,
and errors to the project design. This is still
currently incomplete.